(*slime-features*): Variable remembering the FEATURES list.
(sync-state-to-emacs): Update Emacs about any state changes -
currently this just means changes to the FEATURES list.

(eval-string): Call `sync-state-to-emacs' before sending result.

(eval-region): With optional PACKAGE-UPDATE-P, if the evaluation
changes the current package, tell Emacs about the new package.

(listener-eval): Tell `eval-region' to notify Emacs of package
changes, so that e.g. (in-package :swank) does the right thing when
evaluated in the REPL.