(slime-face-attributes, slime-face-font-name): Copy the font too.
(slime-buffer-package): Try to find be the package name before
resorting to slime-buffer-package.  Return nil and not "CL-USER" if
the package cannot be determined.

(slime-goto-location): Insert notes with a source path but no
filename or buffername at point.  This can happen for warnings during
macro expansion. (The macro expander is a interpreted function and
does not have a filename or buffername.)

(slime-show-note): Display 2 double quotes "" in the for zero length
messages.  SERIES tends to signal warnings with zero length messages.

slime-complete-saved-window-configuration: Store the window config in
a buffer local variable.

(slime-print-apropos): Add support for alien types.

(slime-select-function): Bind pop-up-windows to nil. (Doesn't seem to
work, though.)

(slime-selector): Switch to the minibuffer for reading the event.

(slime-display-buffer-region): Enlarge the window if it is too small.

(slime-find-buffer-package): Initialize command hooks.