(listener-eval): Format results in *buffer-package*.
Exporting (CREATE-SWANK-SERVER <port>). This function can be
called directly to start a swank server, which you can then
connect to with `M-x slime-connect'. It takes a port number as
argument, but this can be zero to use a random available port.
The function always returns the actual port number being used.