Some tweaking to the REPL.  slime-repl-input-end-mark is now always left inserting and slime-mark-input-end "deactivates" the end mark by moving it the beginning of the buffer.
(slime-goto-source-location): Next try for more uniform
source-locations.  A source-location is now a structure with a
"buffer-designator" and  "position-designator".  The buffer-designator
open the file or buffer and the position-designator moves point to the
right position.


(slime-autodoc-mode): New command.

(slime-find-fdefinitions): Experimental support for generic functions
with methods.
(slime-show-xrefs, slime-insert-xrefs, slime-goto-xref): Rewritten to
work with more general source locations.