(connection): Added slots to track indentation caching.
(*connections*): List of all open connections.  (default-connection):
Function to get a "default" connection. This is intended to support
globally using the debugger hook outside the context of a SLIME
request, which is broken at present.

(with-connection): Don't setup a restart: that must be done
separately.

(sync-state-to-emacs): Call `update-connection-indentation'.

(update-connection-indentation): Automatically discover how to indent
macros and tell Emacs.