(emacs-connected, make-stream-interactive): Move the soft-force-output stuff to make-stream-interactive.
(frame-source-location-for-emacs): Pass the function name of the next
(newer) frame as a hint to Emacs.  This way we can highlight the call
site in some cases, instead of the entire defun.

(frame-location): Renamed from function-name-location.  The argument
is now a dspec, not only a name. Also include hints for Emacs.

(lispworks-inspect): Simplified from old code.
(inspect-for-emacs): Use it for also for simple functions.